create-pull-request.py: Do not update already existing pull request.
Updating an existing pull request is tricky: Force pushing a commit
will always update the PR. But we only would want to do this if there
are actual changes. In addition to that we would need to post the new
log as comment. However it's easier to just bail out. So let's do that
for now.