LearnMoreSwitchPreference: Let Android add the switch widget. (#1830)
The previous version of the code added a switch with the id "android:id/switch_widget".
This ID doesn't seem to be used by earlier versions of Android. The easiest way to fix
this is to add the "widget_frame" id to the parent layout and let Android add the
widget with whatever ID it needs.